Improve TypeScript support for Exact Optional Property Types.
Downstream projects which have set exactOptionalPropertyTypes to true
will get compile errors when providing values that might be undefined.

When this flag is set, TypeScript will treat differently the following
types:

    { a?: string }      and
    { a: string | undefined }

The former will accept {} and { a: 'test' } as valid values but
reject { a: undefined } while the latter will reject {} but accept
{ a: undefined }.

Given that we want to accept both in most cases, it's best to use
where appropriate:

    { a?: string | undefined }

This works well under the new flag and is backward compatible with
older version of TypeScript and projects that haven't adopted the
new flag.

Note:

  The previous examples use `string` only because it helps reasoning
  about the underlying issue. Any other types other than undefined
  would have the same problem.